app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
jinja2==3.1.4
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
apscheduler==3.10.4
passlib==1.7.4
PyJWT==2.9.0
bcrypt==3.2.2
APScheduler==3.10.4
pytz==2024.1
